[mesh][metrics] finalize metrics subsystem cleanup
- Move the MeshMetrics recorder implementation into the metrics subsystem and remove legacy metrics shims
- Centralize metric names and path normalization through schema helpers
- Remove the core metrics aggregator shim
