--- atom/mesh/src/core/metrics_aggregator.rs
@@ -1,1 +1,0 @@
-pub use crate::observability::metrics::engine_metrics::{MetricPack, aggregate_metrics};
