+++ atom/mesh/src/observability/inflight_tracker.rs
@@ -8,13 +8,16 @@ use std::{
 
 use dashmap::DashMap;
 
-use super::gauge_histogram::{BucketBounds, GaugeHistogramHandle, GaugeHistogramVec};
+use super::{
+    gauge_histogram::{BucketBounds, GaugeHistogramHandle, GaugeHistogramVec},
+    metrics::schema::names,
+};
 use crate::policies::utils::PeriodicTask;
 
 static INFLIGHT_AGE_BOUNDS: BucketBounds<11> =
     BucketBounds::new([30, 60, 180, 300, 600, 1200, 3600, 7200, 14400, 28800, 86400]);
 static INFLIGHT_AGE_HISTOGRAM: GaugeHistogramVec<11> =
-    GaugeHistogramVec::new("mesh_http_inflight_request_age_count", &INFLIGHT_AGE_BOUNDS);
+    GaugeHistogramVec::new(names::HTTP_INFLIGHT_REQUEST_AGE_COUNT, &INFLIGHT_AGE_BOUNDS);
 static INFLIGHT_AGE_HANDLE: LazyLock<GaugeHistogramHandle> =
     LazyLock::new(|| INFLIGHT_AGE_HISTOGRAM.register_no_labels());
