+++ atom/mesh/src/observability/metrics/mod.rs
@@ -1,3 +1,10 @@
+//! Metrics subsystem entry point.
+//!
+//! Business code should depend on the re-exported facade and helpers here rather
+//! than reaching into individual modules for metric names or implementation
+//! details. Engine metrics aggregation and Mesh self metrics exposure stay in
+//! separate submodules to keep their scrape paths and failure semantics distinct.
+
 pub mod config;
 pub mod engine_metrics;
 pub mod mesh_metrics;
@@ -5,15 +12,12 @@ pub mod recorder;
 pub mod routes;
 pub mod schema;
 
-#[path = "../metrics.rs"]
-mod legacy;
-
 pub use config::PrometheusConfig;
-pub use recorder::{MeshMetrics, Metrics, StreamingMetricsParams};
+pub use recorder::{MeshMetrics, StreamingMetricsParams};
 pub use routes::MetricsRouteFactory;
 pub use schema::{
     METRIC_INVENTORY, MetricKind, MetricSpec, MetricStatus, bool_to_static_str,
-    labels as metrics_labels, method_to_static_str, normalize_path, status_code_to_cow,
+    labels as metrics_labels, method_to_static_str, normalize_path_for_metrics, status_code_to_cow,
     status_code_to_static_str,
 };
