+++ atom/mesh/src/observability/mod.rs
@@ -4,5 +4,4 @@ pub mod events;
 pub mod gauge_histogram;
 pub mod inflight_tracker;
 pub mod logging;
-#[path = "metrics/mod.rs"]
 pub mod metrics;
